# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Station API",
    "version": "0.1.0",
    "description": "Generated OpenAPI for the first-pass Station portability route set."
  },
  "paths": {
...
    "/api/registry/plugins": {
      "get": {
        "operationId": "listRegistryPlugins",
        "summary": "List available registry plugins",
        "tags": [
          "registry"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation or request error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/registry/plugins/{id}": {
      "delete": {
        "operationId": "uninstallRegistryPlugin",
        "summary": "Uninstall a registry plugin",
        "tags": [
          "registry"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Removed; runtime activation, withdrawn plugin command effects still winding down, or a withdrawal that could not be recorded (commandEffectsUnavailable)"
          },
          "400": {
            "description": "Validation or request error"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
    "/api/plugins": {
      "get": {
        "operationId": "listPlugins",
        "summary": "List installed plugins",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Validation or request error"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/plugins/{name}": {
      "delete": {
        "operationId": "deletePlugin",
        "summary": "Delete an installed plugin",
        "tags": [
          "plugins"
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "202": {
            "description": "Removed; runtime activation, withdrawn plugin command effects still winding down, or a withdrawal that could not be recorded (commandEffectsUnavailable)"
          },
          "400": {
            "description": "Validation or request error"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "parameters": [
        {
          "name": "name",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
  },
...
}